Changes A_path.c

+++ A_path.c
@@ -1,12 +1,12 @@
 #include "shell.h"
 
 /**
-* finding_path - this function finds the path from the global env
-* Return: returns NULL if path is not found or path if path is found.
+* finding_path - A function finds the path from the global env.
+* Return: returns NULL if path not found or path if path is found.
 */
 char *finding_path(void)
 {
-	int z;
+	size_t z;
 	char **enve = environ, *apath = NULL;
 
 	while (*enve)
